Additive inverse for different integer types

+++ etc/t1.c
@@ -1,4 +1,4 @@
-int fun(int a, int b, int c, int d, int e, int f, int g, int h)
+long fun(long a, long b, long c, long d, long e, long f, long g, long h)
 {
   return a * b + c * d + e * f + g * h;
 }
